Processor/controller timer calculators with prescalers
Calculators to determine tick counts and overflows for timers when using interrupts and not. 
Works for any register size and any prescaler. Originally written for 8-bit AVR, but will likely work with anything with a fixed clock speed.
Note: when using the nice calculator, the clock speed entered in MHz as integers, so if that doesn't fit your clock, you'll have to change/remove the multiplier and use Hz or kHz.